# docs/swagger.yaml
definitions:
...
  personalfinancedss_internal_module_cashflow_budget_domain.AlertThreshold:
    enum:
    - "50"
    - "75"
    - "90"
    - "100"
    type: string
    x-enum-comments:
      AlertAt50: Alert at 50% spent
      AlertAt75: Alert at 75% spent
      AlertAt90: Alert at 90% spent
      AlertAt100: Alert when exceeded
    x-enum-descriptions:
    - Alert at 50% spent
    - Alert at 75% spent
    - Alert at 90% spent
    - Alert when exceeded
    x-enum-varnames:
    - AlertAt50
    - AlertAt75
    - AlertAt90
    - AlertAt100
  personalfinancedss_internal_module_cashflow_budget_domain.BudgetPeriod:
    enum:
    - daily
    - weekly
    - monthly
    - quarterly
    - yearly
    - custom
    type: string
    x-enum-varnames:
    - BudgetPeriodDaily
    - BudgetPeriodWeekly
    - BudgetPeriodMonthly
    - BudgetPeriodQuarterly
    - BudgetPeriodYearly
    - BudgetPeriodCustom
  personalfinancedss_internal_module_cashflow_budget_domain.BudgetStatus:
    enum:
    - active
    - exceeded
    - warning
    - paused
    - expired
    type: string
    x-enum-comments:
      BudgetStatusWarning: 80-100% spent
    x-enum-descriptions:
    - ""
    - ""
    - 80-100% spent
    - ""
    - ""
    x-enum-varnames:
    - BudgetStatusActive
    - BudgetStatusExceeded
    - BudgetStatusWarning
    - BudgetStatusPaused
    - BudgetStatusExpired
  personalfinancedss_internal_module_cashflow_budget_dto.BudgetResponse:
    properties:
      account_id:
        type: string
      alert_thresholds:
        items:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_domain.AlertThreshold'
        type: array
      allow_rollover:
        type: boolean
      amount:
        type: number
      auto_adjust:
        type: boolean
      auto_adjust_based_on:
        type: string
      auto_adjust_percentage:
        type: integer
      carry_over_percent:
        type: integer
      category_id:
        type: string
      created_at:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      enable_alerts:
        type: boolean
      end_date:
        type: string
      id:
        type: string
      last_calculated_at:
        type: string
      name:
        type: string
      notification_sent:
        type: boolean
      percentage_spent:
        type: number
      period:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_domain.BudgetPeriod'
      remaining_amount:
        type: number
      rollover_amount:
        type: number
      spent_amount:
        type: number
      start_date:
        type: string
      status:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_domain.BudgetStatus'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  personalfinancedss_internal_module_cashflow_budget_dto.BudgetSummaryResponse:
    properties:
      active_budgets:
        type: integer
      average_percentage:
        type: number
      budgets_by_category:
        additionalProperties:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_service.CategoryBudgetSum'
        type: object
      exceeded_budgets:
        type: integer
      total_amount:
        type: number
      total_budgets:
        type: integer
      total_remaining:
        type: number
      total_spent:
        type: number
      warning_budgets:
        type: integer
    type: object
  personalfinancedss_internal_module_cashflow_budget_dto.CreateBudgetRequest:
    properties:
      account_id:
        type: string
      alert_thresholds:
        items:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_domain.AlertThreshold'
        type: array
      allow_rollover:
        type: boolean
      amount:
        type: number
      auto_adjust:
        type: boolean
      auto_adjust_based_on:
        type: string
      auto_adjust_percentage:
        maximum: 100
        minimum: 0
        type: integer
      carry_over_percent:
        maximum: 100
        minimum: 0
        type: integer
      category_id:
        type: string
      currency:
        type: string
      description:
        type: string
      enable_alerts:
        type: boolean
      end_date:
        type: string
      name:
        type: string
      period:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_domain.BudgetPeriod'
      start_date:
        type: string
    required:
    - amount
    - currency
    - name
    - period
    - start_date
    type: object
  personalfinancedss_internal_module_cashflow_budget_dto.UpdateBudgetRequest:
    properties:
      account_id:
        type: string
      alert_thresholds:
        items:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_domain.AlertThreshold'
        type: array
      allow_rollover:
        type: boolean
      amount:
        type: number
      auto_adjust:
        type: boolean
      auto_adjust_based_on:
        type: string
      auto_adjust_percentage:
        maximum: 100
        minimum: 0
        type: integer
      carry_over_percent:
        maximum: 100
        minimum: 0
        type: integer
      category_id:
        type: string
      currency:
        type: string
      description:
        type: string
      enable_alerts:
        type: boolean
      end_date:
        type: string
      name:
        type: string
      period:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_domain.BudgetPeriod'
      start_date:
        type: string
      status:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_domain.BudgetStatus'
    type: object
...
  personalfinancedss_internal_module_cashflow_budget_service.CategoryBudgetSum:
    properties:
      amount:
        type: number
      category_id:
        type: string
      category_name:
        type: string
      percentage:
        type: number
      remaining:
        type: number
      spent:
        type: number
    type: object
...
  personalfinancedss_internal_shared.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
  personalfinancedss_internal_shared.Success:
    properties:
      detail:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact:
    email: support@personalfinancedss.com
    name: API Support
    url: http://www.swagger.io/support
  description: Personal Finance Decision Support System API for managing accounts,
    transactions, budgets, goals, and investments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Personal Finance DSS API
  version: "1.0"
paths:
...
  /api/v1/budgets:
    get:
      description: Get all budgets for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_dto.BudgetResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all user budgets
      tags:
      - budgets
    post:
      consumes:
      - application/json
      description: Create a new budget for the authenticated user
      parameters:
      - description: Budget details
        in: body
        name: budget
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_dto.CreateBudgetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_dto.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new budget
      tags:
      - budgets
  /api/v1/budgets/{id}:
    delete:
      description: Delete a budget
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete budget
      tags:
      - budgets
    get:
      description: Get a specific budget by its ID
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_dto.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get budget by ID
      tags:
      - budgets
    put:
      consumes:
      - application/json
      description: Update an existing budget
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      - description: Budget details
        in: body
        name: budget
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_dto.UpdateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_dto.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update budget
      tags:
      - budgets
  /api/v1/budgets/{id}/recalculate:
    post:
      description: Recalculate spent amount for a budget
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_dto.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Recalculate budget spending
      tags:
      - budgets
  /api/v1/budgets/active:
    get:
      description: Get all active budgets for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_dto.BudgetResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get active budgets
      tags:
      - budgets
  /api/v1/budgets/recalculate-all:
    post:
      description: Recalculate spent amounts for all user budgets
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.Success'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Recalculate all budgets
      tags:
      - budgets
  /api/v1/budgets/summary:
    get:
      description: Get a summary of budget performance for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_budget_dto.BudgetSummaryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get budget summary
      tags:
      - budgets
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme. Enter 'Bearer'
      [space] and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"